+++ lib/src/utils/generator_esc_pos.dart
@@ -321,7 +321,7 @@ class GeneratorEscPos extends Generator {
   }
 
   @override
-  List<int> cut({PosCutMode mode = PosCutMode.full, int extraLines = 5}) {
+  List<int> cut({PosCutMode mode = PosCutMode.full, int extraLines = 2}) {
     var bytes = emptyLines(extraLines);
     switch (mode) {
       case PosCutMode.partial: